feat: move scripts to ~/.local/dotbin
The managed `~/.local/dotbin` directory was created an added to the
path. This enables better separation of managed scripts from those
installed by other software on the target that end up in
`~/.local/bin`. Now it is not possible for legacy scripts to persist on
the system when they are removed fom this repository, as the `exact_`
attribute is given to `dotbin`.